# main/projects/cookbook-api/src/models/nutrition.rs
use serde::Serialize;
use sqlx::FromRow;
use utoipa::ToSchema;

#[derive(Debug, Serialize, FromRow, ToSchema)]
pub struct Nutrition {
    pub id: i64,
    pub recipe_id: i64,
    pub protein_g: Option<f64>,
    pub fat_g: Option<f64>,
    pub carbs_g: Option<f64>,
    pub sodium_mg: Option<f64>,
}