# main/projects/cookbook-web/openapi.json
{"openapi":"3.0.3","info":{"title":"菜谱 Cookbook API","description":"基于 Axum + SQLite 的菜谱查询服务，收录约 700 道中餐菜谱。","contact":{"name":"Cookbook","email":"dev2jake@gmail.com"},"license":{"name":""},"version":"1.0.0"},"paths":{"/api/v1/categories":{"get":{"tags":["分类与统计"],"summary":"获取分类列表","operationId":"list","responses":{"200":{"description":"分类列表（按菜谱数量降序）","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryCount"}}}}}}}},"/api/v1/ingredients":{"get":{"tags":["食材"],"summary":"获取食材列表（按出现次数降序）","operationId":"list","parameters":[{"name":"page","in":"query","description":"页码（从 1 开始）","required":false,"schema":{"type":"integer","format":"int64","default":1,"minimum":1}},{"name":"per_page","in":"query","description":"每页数量（1-100）","required":false,"schema":{"type":"integer","format":"int64","default":50,"maximum":100,"minimum":1}}],"responses":{"200":{"description":"分页食材列表","content":{"application/json":{"schema":{"type":"object","required":["data","page","per_page","total","total_pages"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/T"}},"page":{"type":"integer","format":"int64","description":"当前页码"},"per_page":{"type":"integer","format":"int64","description":"每页数量"},"total":{"type":"integer","format":"int64","description":"总记录数"},"total_pages":{"type":"integer","format":"int64","description":"总页数"}}}}}},"422":{"description":"参数校验失败","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/ingredients/suggest":{"get":{"tags":["食材"],"summary":"食材名模糊搜索（自动补全）","operationId":"suggest","parameters":[{"name":"q","in":"query","description":"搜索关键词（食材名模糊匹配）","required":false,"schema":{"type":"string","nullable":true}},{"name":"limit","in":"query","description":"返回条数上限","required":false,"schema":{"type":"integer","format":"int64","default":10,"nullable":true,"maximum":50,"minimum":1}}],"responses":{"200":{"description":"匹配的食材列表","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IngredientSummary"}}}}}}}},"/api/v1/meal-plan":{"post":{"tags":["餐单"],"summary":"生成每日餐单","operationId":"generate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MealPlanRequest"}}},"required":true},"responses":{"200":{"description":"生成的餐单","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MealPlanResponse"}}}},"422":{"description":"参数校验失败","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/recipes":{"get":{"tags":["菜谱"],"summary":"获取菜谱列表（支持分页、过滤、排序）","operationId":"list","parameters":[{"name":"page","in":"query","description":"页码（默认 1）","required":false,"schema":{"type":"integer","format":"int64","default":1,"minimum":1}},{"name":"per_page","in":"query","description":"每页数量（默认 20，上限 100）","required":false,"schema":{"type":"integer","format":"int64","default":20,"maximum":100,"minimum":1}},{"name":"category","in":"query","description":"分类过滤（如 水产、早餐）","required":false,"schema":{"type":"string","nullable":true}},{"name":"difficulty_min","in":"query","description":"难度下限（1-5）","required":false,"schema":{"type":"integer","format":"int64","nullable":true}},{"name":"difficulty_max","in":"query","description":"难度上限（1-5）","required":false,"schema":{"type":"integer","format":"int64","nullable":true}},{"name":"min_calories","in":"query","description":"热量下限（大卡）","required":false,"schema":{"type":"number","format":"double","nullable":true}},{"name":"max_calories","in":"query","description":"热量上限（大卡）","required":false,"schema":{"type":"number","format":"double","nullable":true}},{"name":"source","in":"query","description":"来源过滤（HowToCook / CookLikeHOC）","required":false,"schema":{"type":"string","nullable":true}},{"name":"sort_by","in":"query","description":"排序字段（calories / difficulty / name，默认 id）","required":false,"schema":{"type":"string","nullable":true}},{"name":"order","in":"query","description":"排序方向（asc / desc，默认 asc）","required":false,"schema":{"type":"string","nullable":true}}],"responses":{"200":{"description":"分页菜谱列表","content":{"application/json":{"schema":{"type":"object","required":["data","page","per_page","total","total_pages"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/T"}},"page":{"type":"integer","format":"int64","description":"当前页码"},"per_page":{"type":"integer","format":"int64","description":"每页数量"},"total":{"type":"integer","format":"int64","description":"总记录数"},"total_pages":{"type":"integer","format":"int64","description":"总页数"}}}}}},"422":{"description":"参数校验失败","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/recipes/by-ingredients":{"get":{"tags":["菜谱"],"summary":"按食材反查菜谱","operationId":"by_ingredients","parameters":[{"name":"ingredients","in":"query","description":"逗号分隔的食材名（支持模糊匹配，如 \"豆腐,鸡蛋\"）","required":false,"schema":{"type":"string","nullable":true}},{"name":"match","in":"query","description":"匹配模式：any（含其中之一）/ all（全部包含），默认 any","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int64","default":1,"minimum":1}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","format":"int64","default":20,"maximum":100,"minimum":1}}],"responses":{"200":{"description":"含指定食材的菜谱列表","content":{"application/json":{"schema":{"type":"object","required":["data","page","per_page","total","total_pages"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/T"}},"page":{"type":"integer","format":"int64","description":"当前页码"},"per_page":{"type":"integer","format":"int64","description":"每页数量"},"total":{"type":"integer","format":"int64","description":"总记录数"},"total_pages":{"type":"integer","format":"int64","description":"总页数"}}}}}},"422":{"description":"参数校验失败","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/recipes/random":{"get":{"tags":["菜谱"],"summary":"随机获取一道菜谱","operationId":"random","parameters":[{"name":"category","in":"query","description":"分类过滤（可选）","required":false,"schema":{"type":"string","nullable":true}},{"name":"max_calories","in":"query","description":"热量上限（可选）","required":false,"schema":{"type":"number","format":"double","nullable":true}}],"responses":{"200":{"description":"随机菜谱详情","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeDetail"}}}},"404":{"description":"没有符合条件的菜谱","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/recipes/search":{"get":{"tags":["菜谱"],"summary":"全文搜索菜谱（FTS5 trigram）","operationId":"search","parameters":[{"name":"q","in":"query","description":"查询词（必填）。≥3 字符使用 FTS5，<3 字符自动 fallback 到菜名 LIKE 匹配","required":false,"schema":{"type":"string","nullable":true}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int64","default":1,"minimum":1}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","format":"int64","default":20,"maximum":100,"minimum":1}}],"responses":{"200":{"description":"搜索结果（按相关度排序）","content":{"application/json":{"schema":{"type":"object","required":["data","page","per_page","total","total_pages"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/T"}},"page":{"type":"integer","format":"int64","description":"当前页码"},"per_page":{"type":"integer","format":"int64","description":"每页数量"},"total":{"type":"integer","format":"int64","description":"总记录数"},"total_pages":{"type":"integer","format":"int64","description":"总页数"}}}}}},"422":{"description":"参数校验失败","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/recipes/{id}":{"get":{"tags":["菜谱"],"summary":"获取菜谱详情（含食材、步骤、营养、标签）","operationId":"get","parameters":[{"name":"id","in":"path","description":"菜谱 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"菜谱详情","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeDetail"}}}},"404":{"description":"菜谱不存在","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/recipes/{id}/similar":{"get":{"tags":["菜谱"],"summary":"获取相似菜谱（同分类，按食材重合度排序）","operationId":"similar","parameters":[{"name":"id","in":"path","description":"菜谱 ID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"limit","in":"query","description":"返回条数（默认 5，上限 20）","required":false,"schema":{"type":"integer","format":"int64","default":5,"maximum":20,"minimum":1}}],"responses":{"200":{"description":"相似菜谱列表","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RecipeSummary"}}}}},"404":{"description":"菜谱不存在","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/stats":{"get":{"tags":["分类与统计"],"summary":"获取全局统计数据","operationId":"stats","responses":{"200":{"description":"全局统计","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatsResponse"}}}},"500":{"description":"服务器错误","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"CategoryCount":{"type":"object","required":["name","count"],"properties":{"count":{"type":"integer","format":"int64","description":"该分类的菜谱数量"},"name":{"type":"string","description":"分类名"}}},"DayPlan":{"type":"object","description":"单日餐单","required":["day","breakfast","lunch","dinner"],"properties":{"breakfast":{"$ref":"#/components/schemas/RecipeSummary"},"day":{"type":"integer","description":"第几天（从 1 开始）","minimum":0},"dinner":{"type":"array","items":{"$ref":"#/components/schemas/RecipeSummary"},"description":"晚餐（2 道）"},"lunch":{"type":"array","items":{"$ref":"#/components/schemas/RecipeSummary"},"description":"午餐（2 道）"}}},"ErrorResponse":{"type":"object","description":"统一错误响应格式","required":["error","message"],"properties":{"error":{"type":"string","description":"错误码"},"message":{"type":"string","description":"错误描述"}}},"Ingredient":{"type":"object","required":["id","recipe_id","name"],"properties":{"amount":{"type":"string","nullable":true},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"note":{"type":"string","nullable":true},"recipe_id":{"type":"integer","format":"int64"},"unit":{"type":"string","nullable":true}}},"IngredientSummary":{"type":"object","required":["name","recipe_count"],"properties":{"name":{"type":"string","description":"食材名称"},"recipe_count":{"type":"integer","format":"int64","description":"出现在多少道菜谱中"}}},"MealPlanRequest":{"type":"object","description":"餐单生成请求","properties":{"days":{"type":"integer","description":"天数（1-14，默认 7）","minimum":0},"max_calories_per_meal":{"type":"number","format":"double","description":"每道菜热量上限（大卡）","nullable":true},"max_difficulty":{"type":"integer","format":"int64","description":"难度上限（1-5）","nullable":true},"people":{"type":"integer","format":"int32","description":"用餐人数（仅记录，默认 2）","minimum":0},"tags":{"type":"array","items":{"type":"string"},"description":"要求同时具有的标签（如 \"低卡\"、\"新手友好\"）","nullable":true}}},"MealPlanResponse":{"type":"object","description":"餐单响应","required":["days","people"],"properties":{"days":{"type":"array","items":{"$ref":"#/components/schemas/DayPlan"}},"people":{"type":"integer","format":"int32","minimum":0}}},"Nutrition":{"type":"object","required":["id","recipe_id"],"properties":{"carbs_g":{"type":"number","format":"double","nullable":true},"fat_g":{"type":"number","format":"double","nullable":true},"id":{"type":"integer","format":"int64"},"protein_g":{"type":"number","format":"double","nullable":true},"recipe_id":{"type":"integer","format":"int64"},"sodium_mg":{"type":"number","format":"double","nullable":true}}},"PagedResult":{"type":"object","required":["data","page","per_page","total","total_pages"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/T"}},"page":{"type":"integer","format":"int64","description":"当前页码"},"per_page":{"type":"integer","format":"int64","description":"每页数量"},"total":{"type":"integer","format":"int64","description":"总记录数"},"total_pages":{"type":"integer","format":"int64","description":"总页数"}}},"RecipeDetail":{"allOf":[{"$ref":"#/components/schemas/RecipeRow"},{"type":"object","required":["ingredients","steps","tags"],"properties":{"ingredients":{"type":"array","items":{"$ref":"#/components/schemas/Ingredient"}},"nutrition":{"allOf":[{"$ref":"#/components/schemas/Nutrition"}],"nullable":true},"steps":{"type":"array","items":{"$ref":"#/components/schemas/Step"}},"tags":{"type":"array","items":{"type":"string"}}}}]},"RecipeRow":{"type":"object","required":["id","name","category","source","source_path","created_at"],"properties":{"calories":{"type":"number","format":"double","nullable":true},"category":{"type":"string"},"cover_image":{"type":"string","nullable":true},"created_at":{"type":"string"},"description":{"type":"string","nullable":true},"difficulty":{"type":"integer","format":"int64","nullable":true},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"source":{"type":"string"},"source_path":{"type":"string"}}},"RecipeSummary":{"type":"object","required":["id","name","category","source","ingredient_count"],"properties":{"calories":{"type":"number","format":"double","nullable":true},"category":{"type":"string"},"cover_image":{"type":"string","nullable":true},"difficulty":{"type":"integer","format":"int64","nullable":true},"id":{"type":"integer","format":"int64"},"ingredient_count":{"type":"integer","format":"int64"},"name":{"type":"string"},"source":{"type":"string"}}},"StatsResponse":{"type":"object","required":["total_recipes","by_category","sources"],"properties":{"avg_calories":{"type":"number","format":"double","description":"平均热量（大卡）","nullable":true},"by_category":{"type":"object","description":"按分类的菜谱数量","additionalProperties":{"type":"integer","format":"int64"}},"sources":{"type":"object","description":"按数据来源的菜谱数量","additionalProperties":{"type":"integer","format":"int64"}},"total_recipes":{"type":"integer","format":"int64","description":"菜谱总数"}}},"Step":{"type":"object","required":["id","recipe_id","step_order"],"properties":{"content":{"type":"string","nullable":true},"id":{"type":"integer","format":"int64"},"image_url":{"type":"string","nullable":true},"recipe_id":{"type":"integer","format":"int64"},"step_order":{"type":"integer","format":"int64"}}}}},"tags":[{"name":"菜谱","description":"菜谱查询、搜索、相似推荐"},{"name":"分类与统计","description":"分类列表和全局统计"},{"name":"食材","description":"食材列表和模糊搜索"},{"name":"餐单","description":"自动生成每日餐单"}]}